Fix TIFF vertical flip, lazy source materialization, and large-image overflows
TIFF orientation (double-flip bug): libtiff's RGBA readers already return a
top-down raster for TOPLEFT images, but the code flipped/reversed the rows
again, producing bottom-up output. Removed the redundant flip in three places:
- TIFFWrapper.cpp (non-lazy, pre-existing): TIFFReadRGBAImageOriented(TOPLEFT)
- TiffLazySource.cpp DecodeSingleStrip: TIFFReadRGBAStrip
- TiffLazySource.cpp DecodeTile: TIFFReadRGBATile
Verified with a libtiff-linked test harness against multistrip and tiled TIFFs.

Lazy source: factor full-decode-once into MaterializeLazySource() and call it
consistently from Resample/GetDIBInternal trapezoid/ConvertSrcTo4Channels/
CreateThumbnailImage so NULL pixel pointers no longer reach the resamplers.

Overflow/UB fixes: 64-bit byte/offset math in PSD, TGA, TiffLazySource,
ColorPalette, ExtractFrames; split unsequenced read-modify-write expressions
in PSDWrapper and ReaderTGA.

Correctness: route PLANARCONFIG_SEPARATE TIFFs through libtiff RGBA reader;
pass row (not strip index) to TIFFReadRGBAStrip; keep sequential decoder
caches (WebP/JXL/APNG) alive across the frame-extraction loop; free AVIF
pixels with delete[] to match the new[] allocator; init m_bTerminate;
bounds-check truncated EXIF APP1; append extension in batch rename when the
pattern has no {ext} token; free pPixelData on TurboJpeg OOM.

+++ src/JPEGView/ColorPalette.cpp
@@ -54,8 +54,10 @@ std::vector<CColorPalette::SColor> CColorPalette::Extract(const void* pPixelsBGR
 	const PixelBGRA* pSrc = static_cast<const PixelBGRA*>(pPixelsBGRA);
 
 	// Downsample so we never process more than ~10000 pixels.
-	int nTotalPixels = nWidth * nHeight;
-	int nMaxSamples = 10000;
+	// 64-bit count: nWidth*nHeight overflows a 32-bit int for large images,
+	// which would make the reserve() below allocate a bogus (huge) size.
+	size_t nTotalPixels = (size_t)nWidth * nHeight;
+	size_t nMaxSamples = 10000;
 	int nStep = 1;
 	if (nTotalPixels > nMaxSamples) {
 		nStep = (int)((double)nTotalPixels / nMaxSamples + 0.5);
@@ -64,7 +66,7 @@ std::vector<CColorPalette::SColor> CColorPalette::Extract(const void* pPixelsBGR
 
 	std::vector<PixelBGRA> samples;
 	samples.reserve(nTotalPixels / nStep + 1);
-	for (int i = 0; i < nTotalPixels; i += nStep) {
+	for (size_t i = 0; i < nTotalPixels; i += nStep) {
 		PixelBGRA p = pSrc[i];
 		// Skip fully transparent pixels — they carry no visible color.
 		if (p.a == 0) continue;
